components for switch case
