+++ src/main/java/org/example/Stock.java
@@ -4,30 +4,13 @@ import org.example.interfaces.StockInterface;
 
 public abstract class Stock implements StockInterface {
 
-    private String nome;
-    private double preco;
+
     private int quantidade;
 
-    public Stock(String nome, double preco, int quantidade) {
-        this.nome = nome;
-        this.preco = preco;
+    public Stock() {}
+
+    public Stock(int quantidade) {
         this.quantidade = quantidade;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    public double getPreco() {
-        return preco;
-    }
-
-    public void setPreco(double preco) {
-        this.preco = preco;
     }
 
     public int getQuantidade() {
