+++ src/main/java/org/example/interfaces/StockInterface.java
@@ -1,10 +1,6 @@
 package org.example.interfaces;
 
 public interface StockInterface {
-    String getNome();
-    void setNome(String nome);
-    double getPreco();
-    void setPreco(double preco);
     int getQuantidade();
     void setQuantidade(int quantidade);
 }
